Added prep.sh
This file if for getting the repo ready for submission by deleting
the unnecessary files to make the repo smaller when uploading. These files
are files that are generated over and over each time the program is
compiled, and these files can be safely deleted whenever - as long as
Visual Studio is not running.